done with Debug

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,8 +6,8 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        double quotient = (double) top / (double) bottom;
-        int result = (int) Math.round(quotient);
+        float quotient = (float) top / bottom;
+        int result = Math.round(quotient);
         return result;
     }
